Add aggregator type; refactor types
