+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/GeneralSection.svelte
@@ -10,7 +10,7 @@
 
 <script lang="ts">
   import KeyValue from "$lib/Advisories/CSAFWebview/KeyValue.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
 
   export let vulnerability: Vulnerability;
   const keys: Array<string> = [];
