+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/ID.svelte
@@ -11,7 +11,7 @@
 <script lang="ts">
   import Collapsible from "$lib/Advisories/CSAFWebview/Collapsible.svelte";
   import KeyValue from "$lib/Advisories/CSAFWebview/KeyValue.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
   export let vulnerability: Vulnerability;
   $: vulnerabilityIds = vulnerability.ids ? vulnerability.ids : [];
 </script>
