+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/Vulnerability.svelte
@@ -22,7 +22,7 @@
   import Remediations from "./remediations/Remediations.svelte";
   import Scores from "./scores/Scores.svelte";
   import Threats from "./threats/Threats.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
 
   export let basePath = "";
   export let vulnerability: Vulnerability;
