+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/flags/Flags.svelte
@@ -12,7 +12,7 @@
   import Collapsible from "$lib/Advisories/CSAFWebview/Collapsible.svelte";
   import ValueField from "$lib/Advisories/CSAFWebview/ValueField.svelte";
   import ValueList from "$lib/Advisories/CSAFWebview/ValueList.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
   export let vulnerability: Vulnerability;
 </script>
