+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/involvements/Involvements.svelte
@@ -10,7 +10,7 @@
 
 <script lang="ts">
   import Collapsible from "$lib/Advisories/CSAFWebview/Collapsible.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
   import ValueField from "$lib/Advisories/CSAFWebview/ValueField.svelte";
   export let vulnerability: Vulnerability;
   $: vulnerabilities = vulnerability.involvements ? vulnerability.involvements : [];
