+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/productStatus/ProductStatus.svelte
@@ -11,7 +11,7 @@
 <script lang="ts">
   import Collapsible from "$lib/Advisories/CSAFWebview/Collapsible.svelte";
   import ProductStatusValueList from "./ProductStatusValueList.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
   export let vulnerability: Vulnerability;
   export let basePath = "";
 </script>
