+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/remediations/Remediation.svelte
@@ -11,7 +11,7 @@
 <script lang="ts">
   import KeyValue from "$lib/Advisories/CSAFWebview/KeyValue.svelte";
   import ProductStatusValueList from "../productStatus/ProductStatusValueList.svelte";
-  import type { Remediation } from "$lib/types";
+  import type { Remediation } from "$lib/pmdTypes";
   import ValueField from "$lib/Advisories/CSAFWebview/ValueField.svelte";
   import ValueList from "$lib/Advisories/CSAFWebview/ValueList.svelte";
   export let remediation: Remediation;
