+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/remediations/Remediations.svelte
@@ -11,7 +11,7 @@
 <script lang="ts">
   import Collapsible from "$lib/Advisories/CSAFWebview/Collapsible.svelte";
   import Remediation from "./Remediation.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
   export let basePath = "";
   export let vulnerability: Vulnerability;
 </script>
