+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/scores/Scores.svelte
@@ -13,7 +13,7 @@
   import CVSSV2 from "./CVSSV2.svelte";
   import CVSSV3 from "./CVSSV3.svelte";
   import ProductStatusValueList from "../productStatus/ProductStatusValueList.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
   export let vulnerability: Vulnerability;
 </script>
