+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/threats/Threat.svelte
@@ -11,7 +11,7 @@
 <script lang="ts">
   import KeyValue from "$lib/Advisories/CSAFWebview/KeyValue.svelte";
   import ValueList from "$lib/Advisories/CSAFWebview/ValueList.svelte";
-  import type { Threat } from "$lib/types";
+  import type { Threat } from "$lib/pmdTypes";
   export let threat: Threat;
   const keys: Array<string> = [];
   const values: Array<string> = [];
