+++ client/src/lib/Advisories/CSAFWebview/vulnerabilities/vulnerability/threats/Threats.svelte
@@ -11,7 +11,7 @@
 <script lang="ts">
   import Collapsible from "$lib/Advisories/CSAFWebview/Collapsible.svelte";
   import Threat from "./Threat.svelte";
-  import type { Vulnerability } from "$lib/types";
+  import type { Vulnerability } from "$lib/pmdTypes";
 
   export let vulnerability: Vulnerability;
   $: threats = vulnerability.threats ? vulnerability.threats : [];
